+++ tests/test_ssh_client.py
@@ -66,4 +66,4 @@ def test_ssh_walk(mock_ssh, tmpdir):
     assert (None, 'a.txt') in files
     assert ('b', 'c.txt') in files
     assert ('b', 'd.txt') in files
-    assert ('b/e', 'f.txt') in files
+    assert ('b{}e'.format(os.sep), 'f.txt') in files